+++ tests/pallas/tpu_ops_test.py
@@ -603,8 +603,11 @@ class OpsTest(ptu.PallasTPUTest):
       axis=[-2, -1],
       dtype=[jnp.int32, jnp.int16, jnp.int8],
       mode=["promise_in_bounds", None],
+      wrap_negative_indices=[True, False],
   )
-  def test_dynamic_gather_along_axis(self, idx_shape, axis, dtype, mode):
+  def test_dynamic_gather_along_axis(
+      self, idx_shape, axis, dtype, mode, wrap_negative_indices
+  ):
     if (axis == -2 and not jtu.is_device_tpu_at_least(version=5)) or (
         axis == -1 and not jtu.is_device_tpu_at_least(version=4)
     ):
@@ -612,10 +615,10 @@ class OpsTest(ptu.PallasTPUTest):
     if dtype != jnp.int32:
       if not jtu.is_libtpu_at_least("0.0.46"):
         self.skipTest("Requires libtpu >= 0.0.46")
-      if mode is None:
+      if wrap_negative_indices:
         self.skipTest(
-            "Under None mode, jnp.take_along_axis will cast indices to int32"
-            " causing a type mismatch for dynamic gather"
+            "With wrap_negative_indices=True, jnp.take_along_axis will cast"
+            " indices to int32 causing a type mismatch for dynamic gather"
         )
       if (
           dtype == jnp.int16 or (dtype == jnp.int8 and axis == -2)
@@ -628,7 +631,13 @@ class OpsTest(ptu.PallasTPUTest):
     x_shape[axis] = (8 * 32 // bitwidth, 128)[axis]
 
     def kernel(x, indices, out):
-      out[...] = jnp.take_along_axis(x[...], indices[...], axis, mode=mode)
+      out[...] = jnp.take_along_axis(
+          x[...],
+          indices[...],
+          axis,
+          mode=mode,
+          wrap_negative_indices=wrap_negative_indices,
+      )
 
     x = np.arange(np.prod(x_shape), dtype=dtype).reshape(x_shape)
     idx = jax.random.randint(